Google-Vision/gcloud-function/requirements.txt
functions-framework==3.*
Pillow==11.0.0
google-cloud-vision==3.8.0
google-cloud-storage==2.18.2
numpy==2.0.2